factor admin functionality out of app.py
progress on #27
