+++ app.py
@@ -178,28 +178,5 @@ def search():
 blueprint = Blueprint('images', __name__, static_url_path='/images', static_folder='images/')
 app.register_blueprint(blueprint)
 
-
-@app.route("/admin")
-@is_admin
-def admin():
-    user = User.query.filter_by(id=current_user.id).first()
-    if user.is_admin:
-        return "admin"
-    return "not admin"
-
-
-@app.route("/roles", methods=["GET", "POST"])
-@is_admin
-def roles():
-        if request.method == "POST":
-            user_id = request.form['user_id']
-            user = User.query.filter_by(id=user_id).first()
-            user.is_admin = is_checked(request, "is_admin")
-            user.can_upload = is_checked(request, "can_upload")
-            user.can_view = is_checked(request, "can_view")
-            db.session.add(user)
-            db.session.commit()
-            return render_template("roles_partial.html", user=user)
-        else:
-            all_users = User.query.all()
-            return render_template("roles.html", all_users=all_users)
+import admin
+app.register_blueprint(admin.blueprint, url_prefix='/admin')
